Add automation to update fleet_integrations/*
Each day it will generate updated modules based on the latest
elastic/integrations repo. It will commit and push the changes.

This is kind of hacky because it is possible that changes in
elastic/integrations are not actually released. That would
result in a module generated from newer sources than what is
in the elastic package registry (EPR).