Disable unused input types for policy_template
Streams that are enabled by default need to be disabled. Previously we
disabled streams for the input type that was being used. But streams under
different input types that existed in the policy template were omitted
so they were enabled. This adds an explicit enabled=false for all unused
inputs.

Fixes #7