generator - honor optional dataset name in manifests
The data stream manifest may contain a `dataset` attribute that specifies
the full {data_stream}.{dataset} name. This was not previously accounted for
and only the name of the directory was used to determine the dataset name.

This changes it to read the `dataset:` attribute from the manifest and parse the
{dataset} name so that it can be used instead of the directory name.

Fixes #21